feat: SAC-GPU实现与Batch探测优化
- 添加 sac_benchmark 基准测试工具
- 添加 benchmark_probe_throughput 吞吐量测试
- 实现 MSAC3bit CPU算法
- 实现 batch_probe_manager GPU批量探测
- 更新 GModel/Network/Trail 支持SAC
- 添加 Batch2/Batch3 测试用例
- 添加 SAC/激活策略对比脚本

+++ include/Network.h
@@ -153,7 +153,7 @@ class IntVar {
   void AddValue(const int a);
 
   // Phase 1.1: Trail 域恢复接口
-  void RestoreBitWord(int word_idx, uint32_t bits);
+  void RestoreBitWord(int word_idx, uint64_t bits);
 
   int value(const int idx) const { return vals_[idx]; }
   int size() const;  // Phase 1.1: 移除 level 参数
